Cập nhật thư mục WebApi

+++ WebApi/Repositories/SQLAuthorRepository.cs
@@ -1,6 +1,7 @@
 ﻿using WebApi.Data;
 using WebApi.Models.Domain;
 using WebApi.Models.DTO;
+using WebApi.Models.Validation;
 
 namespace WebApi.Repositories
 {
@@ -13,6 +14,7 @@ namespace WebApi.Repositories
         }
         public AddAuthorRequestDTO AddAuthor(AddAuthorRequestDTO addAuthorRequestDTO)
         {
+
             var author = new Author()
             {
                 FullName = addAuthorRequestDTO.FullName,
@@ -27,6 +29,10 @@ namespace WebApi.Repositories
             var author = _context.Authors.Find(id);
             if (author != null)
             {
+                bool hasBooks = _context.BookAuthors.Any(x => x.AuthorID == id);
+                if (hasBooks)
+                    throw new InvalidOperationException("Author still linked to books. Remove links in Book_Author first.");
+
                 _context.Authors.Remove(author);
                 _context.SaveChanges();
                 return author;
@@ -60,6 +66,9 @@ namespace WebApi.Repositories
 
         public AuthorNoIdDTO UpdateAuthorById(int id, AuthorNoIdDTO authorNoIdDTO)
         {
+            var name = authorNoIdDTO.FullName?.Trim();
+            if (!Validators.MinLength(name, 3))
+                throw new ArgumentException("Author.Name must have at least 3 characters");
             var author = _context.Authors.Find(id);
             if (author != null)
             {
